requirements.txt
pyside6
python-socketio[asyncio_client]
requests
obsws-python